Revert to markov chatbot module to prevent shitty cb.io errors
